+++ src/n64/vr4300/cache.hpp
@@ -9,6 +9,6 @@ namespace n64::vr4300 {
 
 void InitCache();
 template<std::signed_integral Int, MemOp> Int ReadCacheableArea(u32 phys_addr);
-template<size_t access_size, typename... MaskT> void WriteCacheableArea(u32 phys_addr, s64 data, MaskT... mask);
+template<u32 access_size, typename... MaskT> void WriteCacheableArea(u32 phys_addr, s64 data, MaskT... mask);
 
 } // namespace n64::vr4300
